# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API de Gestão de Projetos",
        "description": "Documentação da API para o sistema de gestão de projetos de energia solar",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/equipamentos": {
            "get": {
                "tags": [
                    "Equipamentos"
                ],
                "summary": "Listar todos os equipamentos",
                "description": "Retorna uma lista de todos os equipamentos.",
                "operationId": "86531d216c46f81a7ce9e1e965f1a249",
                "responses": {
                    "200": {
                        "description": "Lista de equipamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Equipamento"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Equipamentos"
                ],
                "summary": "Criar um novo equipamento",
                "description": "Cria um novo equipamento com os dados fornecidos.",
                "operationId": "595f919522a66c73324351568951eec0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "description": "Descrição do equipamento",
                                        "type": "string",
                                        "example": "Inversor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipamento"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar equipamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar equipamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipamentos/{id}": {
            "get": {
                "tags": [
                    "Equipamentos"
                ],
                "summary": "Obter detalhes de um equipamento",
                "description": "Retorna os detalhes de um equipamento específico baseado no ID fornecido.",
                "operationId": "0acfc4624185c75c0e86ccde090e4f23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do equipamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do equipamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipamento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Equipamento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Equipamentos"
                ],
                "summary": "Atualizar um equipamento",
                "description": "Atualiza os dados de um equipamento existente baseado no ID fornecido.",
                "operationId": "081b7d189d4478269bb91da94f9778b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do equipamento a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "description": "Descrição do equipamento",
                                        "type": "string",
                                        "example": "Microinversor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Equipamento atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipamento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Equipamento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar equipamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar equipamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Equipamentos"
                ],
                "summary": "Deletar um equipamento",
                "description": "Deleta um equipamento existente baseado no ID fornecido.",
                "operationId": "6e27b87c4972aa8756543259922708f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do equipamento a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipamento deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Equipamento deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipamento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Equipamento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar equipamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao deletar equipamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Equipamento": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Inversor"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-29T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-29T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}